feat:divided API between Authentication and Public; Added PATCH to update all tables
