+++ srcs/database/docker-compose.yaml
@@ -13,15 +13,11 @@ services:
     volumes:
       - db_data:/var/lib/mysql
       - ./db/init:/docker-entrypoint-initdb.d
-    env_file:
-      - .env
     environment:
       MYSQL_ROOT_PASSWORD:  ${DB_ROOT_PASSWORD}
       MYSQL_DATABASE:       ${DB_NAME}
       MYSQL_USER:           ${DB_USER}
       MYSQL_PASSWORD:       ${DB_PASSWORD}
-    networks:
-      - shared-network
 
   data-service:
     build: ./data-service
@@ -39,8 +35,6 @@ services:
       DB_NAME: ${DB_NAME}
       DB_USER: ${DB_USER}
       DB_PASSWORD: ${DB_PASSWORD}
-    networks:
-      - shared-network
 
 volumes:
   db_data:
